feat: playground cfg dialog ui reorg
